requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
openpyxl
python-dotenv
requests
pydantic
pydantic-settings
